Create security group if none are passed in
If a user does not provide any security_group_ids, then we create one
for them to use.

This security group sets the following rules:
- Consul client ingress rules (port 8301) for the HVN CIDR ranges
- Allows all ports between members of the same security_group
- Allows egress to the Internet

In addition, we fix an issue with the template.sh script that required
interactive input when downloading apt repositories.

We also provide optional variables to expose the EC2 public IPs on port
8080, so that a user can easily access the demo applications.