examples: use remote registry source in examples (#11)
Add functionality in the tests that maps remote module sources to local
module sources, copies the examples to a temp dir, modifies the source
to the local path, and then runs the test. This ensures we will test the
new versions of modules before they are released on the TF Registry.